move archive operation into build template
The issue was not actually the lack of a set executable bit on the
binaries, which was there already. Rather, transferring files through
the artifact pipeline prior to being archived was stripping them of this
data. As a side effect, it's much cleaner to do it in this stage anyhow!